requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.2
Jinja2==3.1.0
MarkupSafe==2.1.1
PyMySQL==1.0.2
python-dotenv==0.20.0
Werkzeug==2.0.3
